Rename email.py -> mail.py to avoid shadowing standard email module
